requirements.txt
flask>=2.0.0
jinja2>=3.0.0
pyyaml>=6.0.0
gunicorn>=20.0.0  # For production deployment 